Move conversation types into their own file
