+++ packages/bbrt/src/drivers/driver-interface.ts
@@ -5,7 +5,7 @@
  */
 
 import type { BBRTChunk } from "../llm/chunk.js";
-import type { BBRTTurn } from "../llm/conversation.js";
+import type { BBRTTurn } from "../llm/conversation-types.js";
 import type { BBRTTool } from "../tools/tool.js";
 import type { Result } from "../util/result.js";
