+++ packages/bbrt/src/drivers/gemini.ts
@@ -5,7 +5,7 @@
  */
 
 import type { BBRTChunk } from "../llm/chunk.js";
-import type { BBRTTurn } from "../llm/conversation.js";
+import type { BBRTTurn } from "../llm/conversation-types.js";
 import type { BBRTTool } from "../tools/tool.js";
 import type { Result } from "../util/result.js";
 import { streamJsonArrayItems } from "../util/stream-json-array-items.js";
