+++ packages/bbrt/src/llm/conversation.ts
@@ -8,69 +8,20 @@ import { Signal } from "signal-polyfill";
 import { SignalArray } from "signal-utils/array";
 import type { SignalSet } from "signal-utils/set";
 import type { BBRTDriver } from "../drivers/driver-interface.js";
-import type { BBRTTool, InvokeResult, ToolInvocation } from "../tools/tool.js";
+import type { BBRTTool, ToolInvocation } from "../tools/tool.js";
 import { BufferedMultiplexStream } from "../util/buffered-multiplex-stream.js";
 import { Lock } from "../util/lock.js";
 import type { Result } from "../util/result.js";
 import { transposeResults } from "../util/transpose-results.js";
 import { waitForState } from "../util/wait-for-state.js";
 import type { BBRTChunk } from "./chunk.js";
-
-export type BBRTTurn = BBRTUserTurn | BBRTModelTurn | BBRTErrorTurn;
-
-export type BBRTUserTurn = BBRTUserTurnContent | BBRTUserTurnToolResponses;
-
-export type BBRTTurnStatus =
-  | "pending"
-  | "streaming"
-  | "using-tools"
-  | "done"
-  | "error";
-
-export interface BBRTUserTurnContent {
-  kind: "user-content";
-  role: "user";
-  status: Signal.State<BBRTTurnStatus>;
-  content: string;
-}
-
-export interface BBRTUserTurnToolResponses {
-  kind: "user-tool-responses";
-  role: "user";
-  status: Signal.State<BBRTTurnStatus>;
-  responses: BBRTToolResponse[];
-}
-
-export interface BBRTModelTurn {
-  kind: "model";
-  role: "model";
-  status: Signal.State<BBRTTurnStatus>;
-  content: AsyncIterable<string>;
-  toolCalls?: SignalArray<BBRTToolCall>;
-  error?: unknown;
-}
-
-export interface BBRTErrorTurn {
-  kind: "error";
-  role: "user" | "model";
-  status: Signal.State<BBRTTurnStatus>;
-  error: unknown;
-}
-
-export interface BBRTToolCall {
-  id: string;
-  tool: BBRTTool;
-  args: unknown;
-  invocation: ToolInvocation;
-}
-
-export interface BBRTToolResponse {
-  id: string;
-  tool: BBRTTool;
-  invocation: ToolInvocation;
-  args: unknown;
-  response: InvokeResult;
-}
+import type {
+  BBRTModelTurn,
+  BBRTToolCall,
+  BBRTToolResponse,
+  BBRTTurn,
+  BBRTTurnStatus,
+} from "./conversation-types.js";
 
 export class BBRTConversation {
   readonly turns = new SignalArray<BBRTTurn>();
